refactor enrolment

+++ src/test/java/ro/myclass/onlineschoolapi/student/rest/StudentResourcesTest.java
@@ -90,7 +90,7 @@ class StudentResourcesTest {
                 .email(faker.internet().emailAddress())
                 .build();
 
-        doReturn(student).when(studentQuerryService).getStudentById((long) 1);
+        doReturn(student).when(studentQuerryService).getStudentById(1);
 
         restMockMvc.perform(get("/api/v1/student/studentById").param("id", String.valueOf(1)))
                 .andExpect(status().isOk())
@@ -99,7 +99,7 @@ class StudentResourcesTest {
 
     @Test
     public void getStudentByIdBadRequest() throws Exception {
-        doThrow(StudentNotFoundException.class).when(studentQuerryService).getStudentById((long) 1);
+        doThrow(StudentNotFoundException.class).when(studentQuerryService).getStudentById( 1);
 
         restMockMvc.perform(get("/api/v1/student/studentById").param("id", String.valueOf(1)))
                 .andExpect(status().isBadRequest());
